requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
opencv-python-headless>=4.8.0
deepface>=0.0.86
tensorflow>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
werkzeug>=3.0.0
tf-keras
huggingface_hub>=0.20.0
